For #175, ensure shrinks of subsequent parms get a chance
When a first parm could be shrunken no more, we correctly detected
this and told the shrink node representing the smallest counterexample
found so far to advance to the next argument to try to shrink it; but
we never generated shrinks from that and gave shrinks of follow-on
parameters a chance to run. This change should correct this issue.